FinRAGFinal/requirements.txt
openai>=1.0.0
anthropic>=0.3.0
tiktoken>=0.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.3
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
tenacity>=8.2.0
PyPDF2>=3.0.0
pandas>=2.0.0
mcp>=1.0.0
mcp-server-stdio>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
torch
transformers
llama-parse
llama-cloud-services
python-dotenv
yfinance
feedparser
rich
pathway
langfuse>=2.0.0
ragas>=0.1.0
